Rename lapack_traits -> lapack
