+++ src/lapack/opnorm.rs
