+++ src/lapack/qr.rs
