+++ src/lapack/solve.rs
