+++ src/lapack/solveh.rs
