+++ src/lapack/triangular.rs
